+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cases where no audio output could be produced before.
 - Significantly improved the performance of playing sound effects
 - Dismantle and Material crafting tabs don't have duplicated recipes
+- Campfires now despawn when underwater
 
 ## [0.10.0] - 2021-06-12